Knoten: - remainingDots abgespeichert - Heuristik nicht abspeichern, stattdessen Aufruf der Heuristikfunktion
IHeuristicFunction: Heuristik von Typ float

Felddistanzen: Mit Lambda wird waehrend der Initialisierung das Maximum ermittelt

+++ src/de/fh/stud/interfaces/IHeuristicFunction.java
@@ -3,5 +3,5 @@ package de.fh.stud.interfaces;
 import de.fh.stud.Suchen.Suchkomponenten.Knoten;
 
 public interface IHeuristicFunction {
-	int calcHeuristic(Knoten node);
+	float calcHeuristic(Knoten node);
 }
